refactor(configuration module)!: Move configuration file functions to module.
This commit only moves the functions, and is thus breaking.
